move channel dependent promotion rules and actions to SyliusCoreBundle
