+++ src/Sylius/Bundle/PromotionBundle/Resources/config/services.xml
@@ -28,9 +28,6 @@
             <argument type="service" id="sylius.promotion_applicator" />
         </service>
 
-        <service id="sylius.promotion_rule_checker.item_total" class="Sylius\Component\Promotion\Checker\Rule\ItemTotalRuleChecker">
-            <tag name="sylius.promotion_rule_checker" type="item_total" label="sylius.form.promotion_rule.item_total" form-type="Sylius\Bundle\PromotionBundle\Form\Type\Rule\ChannelBasedItemTotalConfigurationType" />
-        </service>
         <service id="sylius.promotion_rule_checker.cart_quantity" class="Sylius\Component\Promotion\Checker\Rule\CartQuantityRuleChecker">
             <tag name="sylius.promotion_rule_checker" type="cart_quantity" label="sylius.form.promotion_rule.cart_quantity" form-type="Sylius\Bundle\PromotionBundle\Form\Type\Rule\CartQuantityConfigurationType" />
         </service>
